kernel/.../KReadOptions.java: Added KReadOptions to be used by {KRun,Kast,KProve}Options
